Initial project setup and added user/notification services

+++ src/main/java/za/co/carhire/service/authentication/UserService.java
@@ -6,9 +6,14 @@ import java.util.List;
 
 public interface UserService {
     User save(User user);
+
     User read(Integer userId);
+
     User update(User user);
+
     void delete(Integer userId);
+
     List<User> findAll();
-    List<User> findByIdNumber(Integer idNumber);
+
+    List<User> findByIdNumber(Long idNumber);
 }
